refactor GAS / ADD tracers for asm implementations
This refactors both of these modules to account for their corresponding
zkasm implementations.  For GAS, we have an entirely new implemetnation
replacing the original Corset constraints.  For ADD, we are dropping
legacy support in the LONDON fork for the non-assembly implementation.
