--- arithmetization/src/main/java/net/consensys/linea/zktracer/module/gas/GasOperationComparator.java
@@ -1,25 +1,0 @@
-/*
- * Copyright ConsenSys Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package net.consensys.linea.zktracer.module.gas;
-
-import java.util.Comparator;
-
-public class GasOperationComparator implements Comparator<GasOperation> {
-  @Override
-  public int compare(GasOperation o1, GasOperation o2) {
-    return o1.gasParameters().compareTo(o2.gasParameters());
-  }
-}
